dz4 init commit
