+++ lib/func/app.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+import '../UI/homep.dart';
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      debugShowCheckedModeBanner: false,
+      title: 'Flutter Demo',
+      theme: ThemeData(
+          primarySwatch: Colors.indigo
+      ),
+      home: const VideoApp(),
+    );
+  }
+}
